Drop appversion from search query to get recent extensions
The "appversion" query probably existed to ensure that only
WebExtensions are downloaded. The lower bound was 57 because that is the
first Firefox version without legacy extension support. The upper bound
was 75 because that was in the far future at the time of writing.
Firefox 75 was released in april 2020, which is far in the past now.

AMO has purged all legacy add-ons (i.e.  add-ons whose support range
ends before version 57), so this is not needed any more.