Migrate from API API v3 to AMO API v5
AMO v3 is deprecated and unsupported. This migrates to AMO v5, while
keeping support for (cached) metadata in the AMO API v3 format.

From the perspective of the data format relevant to webextaware,
the main difference is that the file is no longer in
`current_version.files[0]`, but in `current_version.file`.

Another change is the removal of the `current_version.is_webextension`
key (from the AMO v5 API), since AMO does not know anything about
WebExtensions any more.